--- src/bgnbd_expectation.h
@@ -1,9 +1,0 @@
-#ifndef BGNBD_EXPECTATION_HPP
-#define BGNBD_EXPECTATION_HPP
-arma::vec bgnbd_expectation(const double r,
-                              const arma::vec& vAlpha_i,
-                              const arma::vec& vA_i,
-                              const arma::vec& vB_i,
-                              const arma::vec& vT_i);
-
-#endif
